backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
django-filter==25.2
django-cors-headers==4.9.0
django-ratelimit>=4.1.0
django-redis>=5.4.0
hiredis>=2.3.2
pandas>=2.0.0
openpyxl>=3.0.0
python-decouple>=3.8

# Production Dependencies
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
